# internal/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  offers.UpsertOfferInput:
    properties:
      available_qty:
        type: integer
      currency:
        type: string
      delivery_eta_hours:
        type: integer
      display_price:
        type: string
      expiry_date:
        type: string
      is_active:
        type: boolean
      medicine_id:
        type: string
      min_order_qty:
        type: integer
    type: object
...
  response.ErrorEnvelope:
    properties:
      error:
        properties:
          code:
            type: string
          details: {}
          message:
            type: string
        type: object
    type: object
...
info:
  contact: {}
  description: Backend API for PharmaLink B2B marketplace
  title: PharmaLink API
  version: "1.0"
paths:
...
  /manufacturer-requests:
    get:
      parameters:
      - description: status
        in: query
        name: status
        type: string
      - description: limit
        in: query
        name: limit
        type: integer
      - description: cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: List manufacturer requests
      tags:
      - manufacturer
    post:
      consumes:
      - application/json
      parameters:
      - description: manufacturer request payload
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Create manufacturer request (Wholesaler)
      tags:
      - manufacturer
  /manufacturer-requests/{id}/quotes:
    post:
      consumes:
      - application/json
      parameters:
      - description: manufacturer request id
        in: path
        name: id
        required: true
        type: string
      - description: quote payload
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Create manufacturer quote (Manufacturer)
      tags:
      - manufacturer
...
  /offers:
    get:
      parameters:
      - description: medicine id
        in: query
        name: medicine_id
        type: string
      - description: limit
        in: query
        name: limit
        type: integer
      - description: cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      summary: List offers
      tags:
      - offers
  /offers/:
    post:
      consumes:
      - application/json
      parameters:
      - description: offer payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/offers.UpsertOfferInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Create offer (Wholesaler)
      tags:
      - offers
  /offers/{id}:
    patch:
      consumes:
      - application/json
      parameters:
      - description: offer id
        in: path
        name: id
        required: true
        type: string
      - description: offer payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/offers.UpsertOfferInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Update offer (Wholesaler)
      tags:
      - offers
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"